requirements.txt
Django==2.2
django-cleanup==5.1.0
django-crispy-forms==1.10.0
djongo==1.3.4
filetype==1.0.7
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
moviepy==1.0.3
munkres==1.1.4
numpy==1.19.5
opencv-python==4.5.1.48
ffmpeg-python==0.2.0
Pillow==8.1.0
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
sqlparse==0.2.4
tqdm==4.56.0
urllib3==1.26.3

